Consulta de parametros incorporada

+++ DRS_tone_generator/Core/Inc/max2871.h
@@ -111,12 +111,12 @@ typedef struct REGISTER5 {
 
 typedef struct MAX2871 {
 
-	unsigned long long freqOut;
-	unsigned long long freqSumCurrent;
-	unsigned long long freqSumRead;
-	unsigned long long freqSumNew;
-	unsigned long long freqBase;
-	unsigned long long lastFreqSumReadTick;
+	unsigned long freqOut;
+	unsigned long freqSumCurrent;
+	unsigned long freqSumRead;
+	unsigned long freqSumNew;
+	unsigned long freqBase;
+	unsigned long lastFreqSumReadTick;
     bool freqOutUpdate;
 
 	REGISTER0_t register0;
